+++ install/RESOURCES/APPS/SaveLink/AppController.h
@@ -1,0 +1,48 @@
+/* 
+####################################################
+### A G N o S t e p  -  Desktop - by Patrick Cardona
+### (c) 2026 - pcardona34 @ Github
+###
+### Thanks for the GNUstep Developers Community
+### This is Free and Open Source software.
+### Read License in the root directory.
+####################################################
+
+####################################################
+### SaveLink: an Internet Shortcuts Manager
+### AppController.h: Interface
+####################################################
+*/
+ 
+#ifndef _PCAPPPROJ_APPCONTROLLER_H
+#define _PCAPPPROJ_APPCONTROLLER_H
+
+#import <AppKit/AppKit.h>
+// Uncomment if your application is Renaissance-based
+//#import <Renaissance/Renaissance.h>
+
+@interface AppController : NSObject
+{
+  IBOutlet NSWindow *_window;
+}
+
+// Class methods...
++ (void)  initialize;
+
+// Initialization
+- (id) init;
+- (void) dealloc;
+//- (void) awakeFromNib; // Moved to SaveLink Class
+
+// Notification methods...
+- (BOOL) applicationShouldTerminate: (id)sender;
+- (void) applicationWillTerminate: (NSNotification *)aNotif;
+- (BOOL) application: (NSApplication *)application
+	    openFile: (NSString *)fileName;
+
+// Actions...
+- (IBAction) showPrefPanel: (id)sender;
+
+@end
+
+#endif