+++ install/RESOURCES/APPS/SaveLink/AppController.m
@@ -1,0 +1,69 @@
+/* 
+####################################################
+### A G N o S t e p  -  Desktop - by Patrick Cardona
+### (c) 2026 - pcardona34 @ Github
+###
+### Thanks for the GNUstep Developers Community
+### This is Free and Open Source software.
+### Read License in the root directory.
+####################################################
+
+####################################################
+### SaveLink: an Internet Shortcuts Manager
+### AppController.m: Implementation
+####################################################
+*/
+
+#import "AppController.h"
+
+@implementation AppController
+
++ (void) initialize
+{
+  NSMutableDictionary *defaults = [NSMutableDictionary dictionary];
+
+  /*
+   * Register your app's defaults here by adding objects to the
+   * dictionary, eg
+   *
+   * [defaults setObject:anObject forKey:keyForThatObject];
+   *
+   */
+  
+  [[NSUserDefaults standardUserDefaults] registerDefaults: defaults];
+  [[NSUserDefaults standardUserDefaults] synchronize];
+}
+
+- (id) init
+{
+  if ((self = [super init]))
+    {
+    }
+  return self;
+}
+
+- (void) dealloc
+{
+  [super dealloc];
+}
+
+- (BOOL) applicationShouldTerminate: (id)sender
+{
+  return YES;
+}
+
+- (void) applicationWillTerminate: (NSNotification *)aNotif
+{
+}
+
+- (BOOL) application: (NSApplication *)application
+	    openFile: (NSString *)fileName
+{
+  return NO;
+}
+
+- (IBAction) showPrefPanel: (id)sender
+{
+}
+
+@end