+++ install/RESOURCES/APPS/SaveLink/SaveLink_main.m
@@ -1,0 +1,24 @@
+/* 
+####################################################
+### A G N o S t e p  -  Desktop - by Patrick Cardona
+### (c) 2026 - pcardona34 @ Github
+###
+### Thanks for the GNUstep Developers Community
+### This is Free and Open Source software.
+### Read License in the root directory.
+####################################################
+
+####################################################
+### SaveLink: an Internet Shortcuts Manager
+### SaveLink_main.m: main function
+####################################################
+*/
+
+#import <AppKit/AppKit.h>
+#import "SaveLink.h"
+
+int main(int argc, const char *argv[])
+{
+  return NSApplicationMain (argc, argv);
+}
+