+++ install/RESOURCES/APPS/TestCuckoo/Delegate.h
@@ -1,0 +1,33 @@
+/* 
+####################################################
+### A G N o S t e p  -  Desktop - by Patrick Cardona
+### (c) 2026 - pcardona34 @ Github
+###
+### Thanks for the GNUstep Developers Community
+### This is Free and Open Source software.
+### Read License in the root directory.
+####################################################
+
+####################################################
+### TestCuckoo: a little app to test sound...
+### Delegate: Interface
+####################################################
+*/
+
+#import <Foundation/Foundation.h>
+#import <AppKit/AppKit.h>
+
+@interface MyDelegate : NSObject
+{
+  NSWindow *myWindow;
+}
+
+- (void) createMenu;
+- (void) createWindow;
+- (void) applicationWillFinishLaunching: (NSNotification *)notification;
+- (void) applicationDidFinishLaunching: (NSNotification *)notification;
+
+// Actions
+- (void) testCuckoo: (id)sender;
+
+@end