+++ install/RESOURCES/APPS/TestCuckoo/Delegate.m
@@ -1,0 +1,109 @@
+/* 
+####################################################
+### A G N o S t e p  -  Desktop - by Patrick Cardona
+### (c) 2026 - pcardona34 @ Github
+###
+### Thanks for the GNUstep Developers Community
+### This is Free and Open Source software.
+### Read License in the root directory.
+####################################################
+
+####################################################
+### TestCuckoo: a little app to test sound...
+### Delegate: Implementation
+####################################################
+*/
+
+#import "Delegate.h"
+
+@implementation MyDelegate : NSObject
+
+- (void) testCuckoo: (id)sender
+{
+  NSString *soundPath = [[NSBundle mainBundle] pathForResource:@"cuckoo" ofType:@"wav"];
+  NSLog(@"Path: %@", soundPath);
+    
+  NSSound *cuckoo;
+  AUTORELEASE(cuckoo = [[NSSound alloc] initWithContentsOfFile:soundPath byReference:NO]);
+  
+  if (!cuckoo) {
+     NSLog(@"Failed to load sound!");
+  } else {
+     [cuckoo play];
+     NSLog(@"Playing sound...");
+  }
+}
+
+- (void) createMenu
+{
+  NSMenu *menu;
+  NSMenu *infoMenu;
+  NSMenuItem *menuItem;
+  
+  menu = [NSMenu new];
+  infoMenu = [NSMenu new];
+  menuItem = [NSMenuItem new];
+  
+  [infoMenu addItemWithTitle: @"Info Panel..." 
+          action: @selector (orderFrontStandardInfoPanel:) 
+          keyEquivalent: @""];
+  [infoMenu addItemWithTitle: @"Help..." 
+          action: @selector (orderFrontHelpPanel:)
+          keyEquivalent: @"?"];
+  
+  menuItem = [menu addItemWithTitle: @"Info..." 
+                 action: NULL 
+                 keyEquivalent: @""];
+  [menu setSubmenu: infoMenu forItem: menuItem];
+  
+  [menu addItemWithTitle: @"Test Cuckoo"  
+        action: @selector (testCuckoo:)  
+        keyEquivalent: @"t"];
+
+  [menu addItemWithTitle: @"Quit"  
+        action: @selector (terminate:)  
+        keyEquivalent: @"q"];
+
+  [NSApp setMainMenu: menu];
+}
+
+- (void) createWindow
+{
+  NSRect rect;
+  unsigned int styleMask = NSTitledWindowMask 
+                           | NSMiniaturizableWindowMask;
+  NSButton *myButton;
+  NSSize buttonSize;
+
+  myButton = [NSButton new];
+  [myButton setTitle: @"Test of Cuckoo Wav sound!"];
+  [myButton sizeToFit];
+  [myButton setTarget: self];
+  [myButton setAction: @selector (testCuckoo:)];
+
+  buttonSize = [myButton frame].size;
+  rect = NSMakeRect (100, 100, 
+                     buttonSize.width, 
+                     buttonSize.height);
+
+  myWindow = [NSWindow alloc];
+  myWindow = [myWindow initWithContentRect: rect
+                       styleMask: styleMask
+                       backing: NSBackingStoreBuffered
+                       defer: NO];
+  [myWindow setTitle: @"Agnostep Testing"];
+  [myWindow setContentView: myButton];
+}
+
+- (void) applicationWillFinishLaunching: (NSNotification *)notification
+{
+  [self createMenu];
+  [self createWindow];
+}
+
+- (void) applicationDidFinishLaunching: (NSNotification *)notification;
+{
+  [myWindow makeKeyAndOrderFront: nil];
+}
+
+@end